# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NeuroWealth API
  version: 1.0.0
  description: |
    REST API for the NeuroWealth platform — AI-assisted portfolio management
    backed by Stellar smart contracts.

    **Versioning:** All endpoints are served under an explicit version prefix
    (`/api/v1/*`). The legacy unversioned paths (`/api/*`) remain available as
    deprecated aliases and return `Deprecation`/`Sunset` headers. Every response
    includes an `X-API-Version` header. See `docs/api-versioning.md`.

    **Breaking-change policy:** This API follows semantic versioning. Breaking
    changes (removed fields, changed response shapes, new required parameters)
    increment the major version and are announced at least two weeks before release.
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: '3000'
  - url: https://api.neurowealth.app
    description: Production
...
security: []
paths:
...
  /api/v1/telegram:
    get:
      tags: [telegram]
      operationId: telegramWebhookHealth
      summary: Telegram webhook health check
      description: Telegram webhook endpoint health check. Returns a simple text response.
      responses:
        '200':
          description: Webhook is alive
          content:
            text/plain:
              schema:
                type: string
                example: Telegram webhook is alive
    post:
      tags: [telegram]
      operationId: handleTelegramMessage
      summary: Handle incoming Telegram message
      description: |
        Receives incoming Telegram messages via webhook.
        Validates the x-telegram-bot-api-secret-token header. Sends the reply through the Telegram Bot API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: object
                  properties:
                    chat:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Telegram chat ID
                          example: 123456789
                    text:
                      type: string
                      description: Message body
                      example: What is my portfolio?
      responses:
        '200':
          description: Message accepted and sent to Telegram
          content:
            text/plain:
              schema:
                type: string
                example: OK
        '401':
          description: Missing or invalid Telegram webhook secret token
          content:
            text/plain:
              schema:
                type: string
                example: 'Forbidden: invalid Telegram secret token'

  # ── Portfolio ──────────────────────────────────────────────────────────────
...
  /api/v1/sub-accounts:
    post:
      tags: [sub-accounts]
      operationId: createSubAccount
      summary: Create a sub-account relationship
      description: |
        Creates a parent→child sub-account link with an initial permission set.
        The authenticated user becomes the parent. The child must exist and must
        not already be a parent (no chained sub-accounts).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childUserId, permissions]
              properties:
                childUserId:
                  type: string
                  format: uuid
                  description: The child user's ID
                permissions:
                  type: array
                  items:
                    $ref: '#/components/schemas/SubAccountPermission'
                  minItems: 1
                  maxItems: 4
                  description: Initial permission set
            example:
              childUserId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              permissions: [VIEW, DEPOSIT]
      responses:
        '201':
          description: Sub-account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  subAccount:
                    $ref: '#/components/schemas/SubAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Child user not found
        '409':
          description: Sub-account relationship already exists
  /api/v1/sub-accounts/{id}/permissions:
    patch:
      tags: [sub-accounts]
      operationId: updateSubAccountPermissions
      summary: Update sub-account permissions
      description: |
        Replaces the permission set for an existing sub-account. Only the
        parent who owns the sub-account can update permissions.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Sub-account ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissions]
              properties:
                permissions:
                  type: array
                  items:
                    $ref: '#/components/schemas/SubAccountPermission'
                  minItems: 1
                  maxItems: 4
            example:
              permissions: [VIEW, DEPOSIT, WITHDRAW]
      responses:
        '200':
          description: Permissions updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  subAccount:
                    $ref: '#/components/schemas/SubAccount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/sub-accounts/{id}:
    delete:
      tags: [sub-accounts]
      operationId: revokeSubAccount
      summary: Revoke a sub-account
      description: |
        Soft-revokes a sub-account by setting its status to REVOKED. Only the
        parent who owns the sub-account can revoke it. Revocation takes effect
        immediately on the next request. The audit trail is preserved.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Sub-account ID
      responses:
        '200':
          description: Sub-account revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  subAccount:
                    $ref: '#/components/schemas/SubAccount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by `POST /api/v1/auth/verify`.
        Include as `Authorization: Bearer <token>`.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: API-Key
      description: >
        Admin API key. Can be provided either as
        `Authorization: Bearer <token>` or as the legacy
        `X-Admin-Token` header. Issued via `POST /api/v1/admin/keys`.
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: >
        Service-to-service token for internal endpoints (e.g., agent status,
        Prometheus metrics). Alternative auth methods include IP allowlisting
        or admin Bearer token.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Unauthorized
        details:
          description: Optional structured detail (array or object)
          oneOf:
            - type: array
              items: {}
            - type: object

    # ── Sub-Accounts ─────────────────────────────────────────────────────────
    SubAccountPermission:
      type: string
      enum: [VIEW, DEPOSIT, WITHDRAW, MANAGE_STRATEGY]
      description: >
        Granular permission for sub-account delegation. Permissions are
        independent flags — having DEPOSIT does not implicitly grant VIEW.
    SubAccountStatus:
      type: string
      enum: [ACTIVE, REVOKED]
      description: ACTIVE means the sub-account is live; REVOKED means it was soft-deleted.
    SubAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parentUserId:
          type: string
          format: uuid
        childUserId:
          type: string
          format: uuid
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/SubAccountPermission'
        status:
          $ref: '#/components/schemas/SubAccountStatus'
        createdAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
          nullable: true

    # ── Fiat ──────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
...
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Validation error
            details:
              - field: amount
                message: Must be a positive number
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not found
...